152-strikes-actions-new- (#200)
Added the rest of the strikes actions - editing a strike and deleting it

I also added a context to store the traineeId, I hope thats ok. I didn't want to pass it down.
I have a plan to use it on the other pages as well, at some point (if you guys agree to it of course)

Also created a confirmation dialog that can be reused :)