135 - One edit button for all (#205)
- I added a title to each page. Please let me know if that's not the right title :grimacing:
- I made one edit button for all the tabs and removed the individual edit buttons (and cancel) that were on each tab. 
- The loader is only shown on initial loads before data exists 
- I implemented a context for the trainee profile tabs. It's nice in my opnion, because it removed some code from the component. I think that the components are a bit more simplified now.
The data flow is now as follows - On load, the trainee is fetched from the api and stored in the context. All the different fields get their values from the context, and set them there too (when the user changes something).  So when saving the changes - there's a function that compares the original trainee data to the objedct stored in the context. The function that does it is generic and ive never written a generic function before so i hope that it good :grimacing: 
Upon successful patch request, we refetch the trainee from the api. Witth this refetch, the sidebar is being updated too.